add share js component
