+++ src/components/App.js
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import '../styles/App.scss';
 import logo from '../images/woman.png';
 import logoAdalab from '../images/logo-adalab.png';
-import dataApi from '../services/api';
+import Share from './Share';
 
 function App() {
   const [person, setPerson] = useState({
@@ -16,8 +16,6 @@ function App() {
     image:
       'http://www.burrosminiatura.com/wp-content/uploads/2019/08/jenny-L.jpg',
   });
-
-  const [resultUrl, setResultUrl] = useState({});
 
   let paletteClass = '';
   const handleInput = (ev) => {
@@ -46,14 +44,6 @@ function App() {
       linkedin: '',
       github: '',
       palette: '1',
-    });
-  };
-
-  const handleShareBtn = (event) => {
-    event.preventDefault(event);
-    dataApi(person).then((data) => {
-      console.log(data);
-      setResultUrl(data);
     });
   };
 
@@ -335,50 +325,7 @@ function App() {
                 </div>
               </div>
             </fieldset>
-
-            <fieldset className="share">
-              <div className="share__div">
-                <i className="fa-icon fa-solid fa-share-nodes share__div--icon"></i>
-                <legend className="share__div--legend">comparte</legend>
-                <i className="fa fa-shield fa-shield-up share__div--arrow js-arrow js-arrow-share-up collapsed"></i>
-                <i className="fa fa-shield share__div--arrow js-arrow js-arrow-share-down"></i>
-              </div>
-
-              <button
-                className="share__button js-btn-create"
-                onClick={handleShareBtn}
-              >
-                <i className="fa-regular fa-address-card share__button--icon"></i>
-                crear tarjeta
-              </button>
-
-              <div className="share__card hidden js-share-card">
-                <h2 className="share__card--title">
-                  La tarjeta ha sido creada:
-                </h2>
-                <a
-                  className="share__card--url js-link-card"
-                  href={person.success ? person.cardURL : null}
-                  target="_blank"
-                  rel="noreferrer"
-                >
-                  {person.success ? person.cardURL : person.error}
-                </a>
-
-                <div className="share__twitter">
-                  <a
-                    className="share__twitter--twit twitter-share-button js-twitter"
-                    href="#"
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    <i className="fa-brands fa-twitter twitter-share-button__icon"></i>
-                    Compartir en twitter
-                  </a>
-                </div>
-                <p className="line"></p>
-              </div>
-            </fieldset>
+            <Share person={person} />
           </form>
         </section>
       </main>
